fix: Order의 price 자료형 Long으로 수정

+++ src/main/java/com/aipia/demo/order/dto/OrderResponseDto.java
@@ -9,7 +9,7 @@ public class OrderResponseDto {
     private Long orderId;
     private Long memberId;
     private String memberName;
-    private Integer price;
+    private Long price;
 
     public OrderResponseDto(Order order) {
         this.orderId = order.getId();
